requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.1
feedparser>=6.0

# Deduplication and similarity
sentence-transformers>=2.2.0
numpy>=1.24.0

# Database
# sqlite3 è built-in in Python, non serve installarlo

# LLM (scegli uno o entrambi)
openai>=1.0.0
anthropic>=0.7.0

# Utilities
python-dateutil>=2.8.0
# hashlib è built-in, non serve installarlo
urllib3>=2.0.0
markdown>=3.4.0

# Optional: per miglior parsing HTML
html5lib>=1.1

# Frontend mock server
flask>=3.0.0
flask-cors>=4.0.0
